refactor(transcription): move Groq and OpenAI transcription services
